kubent/pyproject.toml
[project]
name = "kubent"
version = "0.1.0-beta"
description = "Kubent: offer an enterprise-level solution to improve agent system better."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "celery[redis]>=5.6.2",
    "docker>=7.1.0",
    "dotenv>=0.9.9",
    "fastapi>=0.124.4",
    "mwin>=0.1.8",
    "openai>=2.11.0",
    "psycopg2>=2.9.11",
    "pyjwt>=2.10.1",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
